requirements.txt
# Causal Impact & Investment Decision Analysis
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyyaml>=6.0

# Causal Inference
pycausalimpact>=0.1.1
statsmodels>=0.14.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Application
streamlit>=1.28.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
reportlab>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest-cov>=4.1.0
hypothesis>=6.82.0
